security: the Go tooling namux and tagtastic already have
uhp-go had no linter, no SAST, no secret scan and no dependency alerting. It
is also the one of the three that brokers credentials and starts subprocesses,
so it is the one that needed them most.

Modelled on namux's .golangci.yml and tagtastic's stated preference for inline
#nosec over blanket excludes, with three deliberate departures:

misspell is not enabled. Both locales are wrong here. US flags 267 lines of
British prose ("recognise", "marshalling"); UK flags "artifact", which is UHP
wire vocabulary, and "initialize", which is an MCP method name. A linter that
cannot be right teaches people to run with -e.

gosec has no excludes at all. namux excludes G404 globally; here every accepted
finding carries an `#nosec <rule> -- reason` at the site instead, so a future
exec or file read nobody has thought about still fails. Ten sites are
annotated: six G304, three G204, one G101.

CI builds with Go 1.24 rather than the 1.22 in go.mod. go.mod states the floor
this module supports; the workflow states what CI builds with, and 1.22 stopped
getting security patches — govulncheck reports 32 stdlib advisories against it.
Raising the toolchain fixes those without raising the floor for consumers.

Three security targets, because they answer three questions. `security` reports
and never fails. `security-push` gates on gosec and gitleaks and reports
govulncheck, since a stdlib advisory is fixed by a toolchain bump rather than by
anything here, and a gate nobody can pass is a gate that gets switched off.
`security-strict` gates on all three, for a release.

Tool versions live in the Makefile and CI reads them, including via
`make gitleaks-version`. That is the same one-number-one-place rule #109 exists
to enforce.

The pre-push hook runs lint and the scanners only when they are installed, so
`make hooks` stays one command rather than a shopping list. CI runs them
unconditionally.

Real code changes are small and separate from the annotations: one redundant
int64 conversion in budget.go, one goimports grouping in inputs.go, and two
//nolint:staticcheck on uhpc's deprecation-window fallbacks, which read the
deprecated field on purpose.

make verify is green: fmt-check, vet, lint, go test -race, the script tests and
both binaries. gitleaks finds nothing across 63 commits.

+++ internal/service/budget.go
@@ -89,7 +89,7 @@ func resolveBudget(requested, configured *int, ceiling time.Duration) time.Durat
 // maxBudgetSeconds is the largest whole number of seconds time.Duration can
 // hold — about 292 years, and so far outside any budget an operator or a client
 // means that the only values above it are overflow attempts and mistakes.
-const maxBudgetSeconds = int64(math.MaxInt64 / int64(time.Second))
+const maxBudgetSeconds = math.MaxInt64 / int64(time.Second)
 
 // resolveStepBudget picks the step budget for one run: the shortest of the
 // three bounds that are set — the request's, the harness's, and the
